pyproject.toml
[project]
name = "llm-cultural-probing"
version = "0.1.0"
description = "Cultural probing of LLMs (India): values/norms, regional variations, kinship."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
dependencies = [
  "jupyterlab",
  "ipykernel",
  "pandas",
  "numpy",
  "matplotlib",
  "seaborn",
  "plotly",
  "pyyaml",
  "tenacity",
  "tqdm",
  "openai",
  "google-generativeai",
  "transformers",
  "torch",
  "accelerate",
  "huggingface-hub",
  "datasets",
  "scikit-learn",
  "tiktoken",
]

[tool.uv]
dev-dependencies = [
  "pytest",
  "ruff",
  "black",
  "nbstripout",
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]


